+++ src/panoramic/cli/cli.py
@@ -201,6 +201,18 @@ def cleanup(target_dataset: str, yes: bool):
     delete_orphans_command(target_dataset=target_dataset, yes=yes)
 
 
+@field_cli.command(help='Scaffold fields defined in models', cls=ContextAwareCommand)
+@click.option('--target-dataset', '-t', type=str, help='Target a specific dataset')
+@click.option('--yes', '-y', is_flag=True, default=False, help='Automatically confirm all actions')
+@handle_exception
+def scaffold(target_dataset: str, yes: bool):
+    from panoramic.cli.command import (
+        scaffold_missing_fields as scaffold_missing_fields_command,
+    )
+
+    scaffold_missing_fields_command(target_dataset=target_dataset, yes=yes)
+
+
 @cli.group()
 def data_connections():
     """Data Connections subcommand for managing data connections.